client/requirements.txt
websocket-client==1.8.0
sounddevice==0.4.7
numpy==1.26.4
pygame==2.6.0
opencv-python==4.10.0.84
Pillow==10.4.0
